# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/cars/autocomplete/{make}": {
            "get": {
                "tags": [
                    "Cars"
                ],
                "summary": "Get all models by car manufacturer",
                "parameters": [
                    {
                        "name": "make",
                        "in": "path",
                        "description": "Parameter for geting all car models",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Ford"
                        },
                        "example": "Frod"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": []
                                        },
                                        "data": {
                                            "required": [
                                                "count",
                                                "items"
                                            ],
                                            "properties": {
                                                "count": {
                                                    "type": "integer",
                                                    "example": 166
                                                },
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1778
                                                            },
                                                            "title": {
                                                                "type": "string",
                                                                "example": "Crown Victoria"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "NOK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "false"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": []
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}